Renamed old party and relationship models, to minimise impact on frontend (for now)
